refactor(storage): extract proximadb-storage-encryption crate
Move src/storage/encryption (AES-256-GCM at-rest encryption: data/file/WAL
layers, HKDF key derivation + rotation, and the transparent EncryptedFilesystem
decorator) into a dedicated workspace crate as a slice of the root-crate
decomposition. The root crate re-exports it as crate::storage::encryption, so
all consumers are untouched.

The EncryptedFilesystem decorator now wraps the FileSystem trait from
proximadb-storage-filesystem-types (extracted in #332) instead of the
root-resident persistence::filesystem module, turning the former upward edge
into a downward crate dependency.

The 4 EncryptedFilesystem decorator tests drive the real root-resident
LocalFileSystem backend (which depends back on encryption, so it cannot be
reached from the leaf crate without a dependency cycle). Rather than introduce
a second in-memory FileSystem implementation, those tests are relocated verbatim
to root (src/storage/encryption_filesystem_tests.rs), reusing LocalFileSystem.
The crate keeps its 14 unit tests.

- root src/ shrinks by ~1.7K LOC across 5 files (one fewer root module)
- proximadb-storage-encryption: 14 unit tests pass, clippy -D warnings clean
- relocated decorator tests: 4 pass from root
- server/bench/migrate binaries build; layering green (storage tier)

Ref ROOT_CRATE_DECOMPOSITION_PLAN_2026_06_21, #332.
