+++ crates/storage/proximadb-storage-encryption/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "proximadb-storage-encryption"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Encryption-at-rest for ProximaDB: AES-256-GCM data/WAL encryption, HKDF key derivation/rotation, and a transparent encrypting FileSystem decorator"
+
+[lib]
+name = "proximadb_storage_encryption"
+path = "src/lib.rs"
+
+[dependencies]
+aes-gcm = "0.10"
+anyhow.workspace = true
+async-trait.workspace = true
+bincode.workspace = true
+hkdf = "0.12"
+memmap2 = "0.9"
+proximadb-storage-filesystem-types.workspace = true
+rand.workspace = true
+rayon.workspace = true
+serde.workspace = true
+sha2 = "0.10"
+tokio.workspace = true
+tracing.workspace = true
+
+[lints]
+workspace = true
