+++ crates/storage/proximadb-storage-encryption/src/file_encryption.rs
@@ -292,7 +292,7 @@ impl FileEncryptionLayer {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::storage::encryption::key_manager::KeyManager;
+    use crate::key_manager::KeyManager;
 
     #[test]
     fn test_file_encryption_round_trip() {
